Rename app/components to app/containers
