+++ src/client/app/Routes.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Route, Switch } from "react-router-dom";
 
-import { Blog, Home, NotFound, Story } from "app/components";
+import { Blog, Home, NotFound, Story } from "app/containers";
 
 interface RoutesInterface {
   Component: React.SFC<any> | typeof React.Component;
