--- src/client/app/components/Story/Story.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-const Story: React.SFC<any> = props => <p>{props.story.value.title}</p>;
-
-export default Story;
