+++ src/client/app/containers/Story/Story.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Story: React.SFC<{ story: any }> = props => (
+  <p>{props.story.value.title}</p>
+);
+
+export default Story;
